requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
Django==4.2.7
django-cors-headers==4.3.0
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.8.0
drf-spectacular==0.29.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mongoengine==0.27.0
mysqlclient==2.2.0
openai==2.15.0
Pillow==10.1.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pymongo==3.12.3
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
requests==2.31.0
rpds-py==0.30.0
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
